refactor: move converter from static class into factory and interface
The current structure is hard coded to only allow for a single
converter based on an underlying wkhtmltopdf binary and Qt browser.

In order to allow for further extension and addition of other
converters based on different underlying webengines this needs to be
moved into a factory based on an interface which may be implemented
by any new converter types.