Issue #23: Prevent race conditions
Race conditions are evident as a result of running multiple
conversions in parallel through adhoc tasks, this results in one or
more conversions failing due to an access key being deleted for a
user when still in use.

Attach an instance ID to access keys and abstract the user key
handling into a key manager, so now only the key for a particular
filename is deleted, not all keys for user.